refactor: use connector link in workflows

+++ packages/cli/src/commands/connector/link.ts
@@ -7,22 +7,45 @@ import { log } from '../../utils.js';
 
 import { getConnectorDirectory, getLocalConnectorPackages, inquireInstancePath } from './utils.js';
 
-const link: CommandModule<{ path?: string }, { path?: string }> = {
+const link: CommandModule<{ path?: string }, { path?: string; cloud: boolean; mock: boolean }> = {
   command: ['link', 'ln'],
   describe: 'Link all connectors in `packages/connectors`, useful for adding local connectors.',
-  handler: async ({ path: inputPath }) => {
+  builder: (yargs) =>
+    yargs
+      .option('cloud', {
+        describe: 'Add additional connectors for Logto Cloud',
+        type: 'boolean',
+        default: false,
+        hidden: true,
+      })
+      .option('mock', {
+        describe: 'Add mock connectors',
+        type: 'boolean',
+        default: false,
+        hidden: true,
+      }),
+  handler: async ({ path: inputPath, cloud, mock }) => {
     const instancePath = await inquireInstancePath(inputPath);
     const packages = await getLocalConnectorPackages(instancePath);
+    const connectorDirectory = getConnectorDirectory(instancePath);
+
+    await fs.mkdir(connectorDirectory, { recursive: true });
+
     await Promise.all(
       packages.map(async ([packageName, packagePath]) => {
+        if (!cloud && packageName.startsWith('connector-logto-')) {
+          return;
+        }
+
+        if (!mock && packageName.startsWith('connector-mock-')) {
+          return;
+        }
+
         try {
-          const targetPath = path.join(
-            getConnectorDirectory(instancePath),
-            '@logto-' + packageName
-          );
+          const targetPath = path.join(connectorDirectory, '@logto-' + packageName);
 
           await fs.rm(targetPath, { recursive: true, force: true });
-          await fs.symlink(path.resolve(packagePath), targetPath);
+          await fs.symlink(path.relative(connectorDirectory, packagePath), targetPath);
         } catch (error) {
           log.warn(error);
